% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNormalityAssessmentApp.R
\name{runNormalityAssessmentApp}
\alias{runNormalityAssessmentApp}
\title{Run the NormalityAssessment Shiny application}
\usage{
runNormalityAssessmentApp()
}
\value{
There is no return value.
}
\description{
Runs the NormalityAssessment Shiny application.
}
\examples{
## only run the app in an interactive R session
if (interactive()) {runNormalityAssessmentApp()}

}
\author{
Christopher Casement \cr
Department of Mathematics \cr
Fairfield University \cr
\email{casementc@gmail.com}

Laura McSweeney \cr
Department of Mathematics \cr
Fairfield University
}
