% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalityAssessment-package.R
\docType{package}
\name{NormalityAssessment-package}
\alias{NormalityAssessment-package}
\alias{_PACKAGE}
\alias{NormalityAssessment}
\title{NormalityAssessment: A Graphical User Interface for Testing Normality Visually}
\description{
The NormalityAssessment package creates plots for assessing
  normality. The methods implemented are based on recent development made in
  graphical inference. In the app, the features in the 'Explore Simulated Data'
  tab enable the user to run the Rorschach procedure, and those in the
  'Include Your Data' tab allow the user to run the line-up procedure.
}
\details{
Package: NormalityAssessment \cr
Type: Package \cr
Version: 0.0.1 \cr
Date: 2021-10-19 \cr
Depends: R (>= 3.5.0) \cr
Imports: dplyr, DT, ggplot2, rio, shiny, shinyalert, shinyBS, stringi,
stringr
License: GPL-3 \cr
BugReports: https://github.com/ccasement/NormalityAssessment/issues \cr
Encoding: UTF-8 \cr
}
\section{Function}{
 \itemize{
\item \code{\link{runNormalityAssessmentApp}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ccasement/NormalityAssessment}
  \item \url{https://CRAN.R-project.org/package=NormalityAssessment}
  \item Report bugs at \url{https://github.com/ccasement/NormalityAssessment/issues}
}

}
\author{
Christopher Casement \cr
Department of Mathematics \cr
Fairfield University \cr
\email{casementc@gmail.com}

Laura McSweeney \cr
Department of Mathematics \cr
Fairfield University
}
